+++ client/include/autd3/gain/primitive.hpp
@@ -3,7 +3,7 @@
 // Created Date: 14/04/2021
 // Author: Shun Suzuki
 // -----
-// Last Modified: 24/09/2021
+// Last Modified: 03/11/2021
 // Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
 // -----
 // Copyright (c) 2021 Hapis Lab. All rights reserved.
@@ -23,7 +23,7 @@ namespace autd::gain {
 
 using core::Gain;
 using core::GainPtr;
-using NullGain = Gain;
+using Null = Gain;
 
 using core::DataArray;
 using core::Vector3;
@@ -157,38 +157,6 @@ class BesselBeam final : public Gain {
   Vector3 _vec_n;
   double _theta_z;
   uint8_t _duty;
-};
-
-/**
- * @brief Gain that can set the phase and duty ratio freely
- */
-class Custom final : public Gain {
- public:
-  /**
-   * @brief Generate function
-   * @param[in] data data of duty ratio and phase of each transducer
-   * @details The data size should be the same as the number of devices you use. The data is 16 bit unsigned integer, where high 8bits represents duty
-   * ratio and low 8bits represents phase.
-   */
-  static GainPtr create(const std::vector<DataArray>& data);
-  /**
-   * @brief Generate function
-   * @param[in] data pointer to data of duty ratio and phase of each transducer
-   * @param[in] data_length length of the data
-   * @details The data length should be the same as the number of transducers you use. The data is 16 bit unsigned integer, where high 8bits
-   * represents duty ratio and low 8bits represents phase
-   */
-  static GainPtr create(const uint16_t* data, size_t data_length);
-  void calc(const core::GeometryPtr& geometry) override;
-  explicit Custom(const std::vector<DataArray>& data) : Gain() { this->_raw_data = data; }
-  ~Custom() override = default;
-  Custom(const Custom& v) noexcept = default;
-  Custom& operator=(const Custom& obj) = default;
-  Custom(Custom&& obj) = default;
-  Custom& operator=(Custom&& obj) = default;
-
- private:
-  std::vector<DataArray> _raw_data;
 };
 
 /**