+++ client/lib/gain/primitive.cpp
@@ -3,7 +3,7 @@
 // Created Date: 14/04/2021
 // Author: Shun Suzuki
 // -----
-// Last Modified: 06/10/2021
+// Last Modified: 08/11/2021
 // Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
 // -----
 // Copyright (c) 2021 Hapis Lab. All rights reserved.
@@ -11,7 +11,6 @@
 
 #include "autd3/gain/primitive.hpp"
 
-#include <algorithm>
 #include <memory>
 #include <vector>
 
@@ -36,7 +35,7 @@ void Grouped::calc(const core::GeometryPtr& geometry) {
   }
 }
 
-GainPtr PlaneWave::create(const Vector3& direction, const double amp) { return create(direction, core::Utilities::to_duty(amp)); }
+GainPtr PlaneWave::create(const Vector3& direction, const double amp) { return create(direction, core::utils::to_duty(amp)); }
 
 GainPtr PlaneWave::create(const Vector3& direction, uint8_t duty) { return std::make_shared<PlaneWave>(direction, duty); }
 
@@ -47,12 +46,12 @@ void PlaneWave::calc(const core::GeometryPtr& geometry) {
   for (size_t dev = 0; dev < geometry->num_devices(); dev++)
     for (size_t i = 0; i < NUM_TRANS_IN_UNIT; i++) {
       const auto dist = geometry->position(dev, i).dot(dir);
-      const auto phase = core::Utilities::to_phase(dist * wavenum);
-      this->_data[dev][i] = core::Utilities::pack_to_u16(this->_duty, phase);
+      const auto phase = core::utils::to_phase(dist * wavenum);
+      this->_data[dev][i] = core::utils::pack_to_u16(this->_duty, phase);
     }
 }
 
-GainPtr FocalPoint::create(const Vector3& point, const double amp) { return create(point, core::Utilities::to_duty(amp)); }
+GainPtr FocalPoint::create(const Vector3& point, const double amp) { return create(point, core::utils::to_duty(amp)); }
 GainPtr FocalPoint::create(const Vector3& point, uint8_t duty) { return std::make_shared<FocalPoint>(point, duty); }
 
 void FocalPoint::calc(const core::GeometryPtr& geometry) {
@@ -60,13 +59,13 @@ void FocalPoint::calc(const core::GeometryPtr& geometry) {
   for (size_t dev = 0; dev < geometry->num_devices(); dev++)
     for (size_t i = 0; i < NUM_TRANS_IN_UNIT; i++) {
       const auto dist = (geometry->position(dev, i) - this->_point).norm();
-      const auto phase = core::Utilities::to_phase(dist * wavenum);
-      this->_data[dev][i] = core::Utilities::pack_to_u16(this->_duty, phase);
+      const auto phase = core::utils::to_phase(dist * wavenum);
+      this->_data[dev][i] = core::utils::pack_to_u16(this->_duty, phase);
     }
 }
 
 GainPtr BesselBeam::create(const Vector3& apex, const Vector3& vec_n, const double theta_z, const double amp) {
-  return create(apex, vec_n, theta_z, core::Utilities::to_duty(amp));
+  return create(apex, vec_n, theta_z, core::utils::to_duty(amp));
 }
 
 GainPtr BesselBeam::create(const Vector3& apex, const Vector3& vec_n, double theta_z, uint8_t duty) {
@@ -86,24 +85,10 @@ void BesselBeam::calc(const core::GeometryPtr& geometry) {
       const auto r = geometry->position(dev, i) - this->_apex;
       const auto rr = rot * r;
       const auto d = std::sin(_theta_z) * std::sqrt(rr.x() * rr.x() + rr.y() * rr.y()) - std::cos(_theta_z) * rr.z();
-      const auto phase = core::Utilities::to_phase(d * wavenum);
-      this->_data[dev][i] = core::Utilities::pack_to_u16(this->_duty, phase);
+      const auto phase = core::utils::to_phase(d * wavenum);
+      this->_data[dev][i] = core::utils::pack_to_u16(this->_duty, phase);
     }
 }
-
-GainPtr Custom::create(const uint16_t* data, const size_t data_length) {
-  const auto dev_num = (data_length + NUM_TRANS_IN_UNIT - 1) / NUM_TRANS_IN_UNIT;
-  std::vector<DataArray> raw_data(dev_num);
-  for (size_t i = 0; i < dev_num; i++) {
-    const auto rem = std::clamp(data_length - i * NUM_TRANS_IN_UNIT, size_t{0}, NUM_TRANS_IN_UNIT);
-    std::memcpy(&raw_data[i][0], data + i * NUM_TRANS_IN_UNIT, rem * sizeof(uint16_t));
-  }
-  return create(raw_data);
-}
-
-GainPtr Custom::create(const std::vector<DataArray>& data) { return std::make_shared<Custom>(data); }
-
-void Custom::calc(const core::GeometryPtr&) { this->_data = std::move(this->_raw_data); }
 
 GainPtr TransducerTest::create(const size_t transducer_index, const uint8_t duty, const uint8_t phase) {
   return std::make_shared<TransducerTest>(transducer_index, duty, phase);
@@ -111,6 +96,6 @@ GainPtr TransducerTest::create(const size_t transducer_index, const uint8_t duty
 
 void TransducerTest::calc(const core::GeometryPtr& geometry) {
   this->_data[geometry->device_idx_for_trans_idx(_transducer_idx)][_transducer_idx % NUM_TRANS_IN_UNIT] =
-      core::Utilities::pack_to_u16(this->_duty, this->_phase);
+      core::utils::pack_to_u16(this->_duty, this->_phase);
 }
 }  // namespace autd::gain